Correcting Claude's code and testing
Corrected several errors in Claude's code, especially around the counter which didn't work properly in his version, but mostly needed things to be moved around and the addition of a single variable for the counter. Tried to test the OSC sending, but I suspect the absence of anywhere to send it might be causing problems (not sure yet if there's an error). The [send] and [pack] objects seem to be working fine - just can't tell if it's actually tried to ship them off. Next step is to set up Unity to receive OSC and see if that helps