ACTUAL (sort of) FUNCTIONAL EXTERNAL!
Now testing in Pd, which is raising some problems. 2025-08-25 18.20_mpe_test_patch.pd is Claude's attempt at a patch for testing, which is so broken it's hilarious. Fortunately, his text instructions are much better. Basically, it seems there's quite a bit of manipulation which needs to happen before getting the data into the new objects - I think I'll need to get him to embed these in the library because otherwise it's just going to be mega confusion and mess. But will suffice for testing that the main part of it works.