Update .gitignore
In my GitHub newbieness, and not having a clue when I started that I would even be using Visual Studio or that it would generate a whole massive bunch of guff, I neglected to switch on the option to ignore said guff. I've switched it on now, but just dealing with the fallout. Hopefully this will make future commits tidier...