Edit code to address errors
Looked like ChattieG just left off the A_FLOAT elements here. I poked about a bit to try to understand what was going on because these were the only 2 errors generated in the code, and this is where I ended up thanks to a combo of the previous lines of code and looking at the declaration of class_addmethod. Don't really have any idea what it actually means, but it seems to have cleared the error.