Simpler version of help for mpe_poly, first steps towards separate voices
Gave ChattieG more info about what I'm doing here and made the point to it that I really don't care 2 figs what channel a note is on - it may be technically necessary for MPE to weave its magic, but it's SO unmusical, so I'm trying to encourage it towards a solution that can be comprehended and used by musicians without their heads imploding. mpe_voice.cpp is intended to create a second 'voice' object to send a single notes and all its associated info to one place so I don't need to wade through a complicated assortment of data just to work out the pitchbend of this one thing. We'll need to make changes to the original mpe_poly.cpp so it feeds the right data through to mpe_voice but I've run out of Chattie credits for now...