Change to route MPE signals for a more musical way of processing
Trying to reshape the outputs so working with the data is less a juggling act dealing with so many channels and more something that's (hopefully) going to be a little more musically intuitive to work with - I figure if I'm making my own library, why does it need to surface a whole bunch of information that just confuses me and which I'm unlikely to need to access directly?
Then needed to change the .def file to add in the new mpe_voice definition.