Tweaking values
It seems my values for the glide might be too low to have any effect in Unity, so added a multiplier to big them up. Also I think this is the first version to add the glide average to the OSC. Needed to add a little pack in there to get both values into the message (something I don't think I'd tried before) but in this version all seems to be going through to Unity well, even if it's not having the effect I want there yet.