Trying to add glide to move the camera
This isn't working yet, but this commit just logs where this is up to at this point where I'm shutting down for the moment to do other things. The forward motion with the view direction (controlled by mouse) is working; and the glide values are being received and recognised by Unity but they just don't seem to be having any effect, which could be because they're too small, but this needs further testing.